+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/hellohttp/extension/StringExtension.kt
@@ -18,3 +18,5 @@ fun String.endWithNewLine() =
     } else {
         this + '\n'
     }
+
+fun String?.countNotBlank(): Int = if (isNullOrBlank()) 0 else 1