+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/hellohttp/ux/local/AppFont.kt
@@ -9,6 +9,7 @@ data class AppFont(
     val buttonFontSize: TextUnit,
     val searchInputSize: TextUnit,
     val supplementSize: TextUnit,
+    val badgeFontSize: TextUnit,
     val codeEditorBodyFontSize: TextUnit,
     val codeEditorLineNumberFontSize: TextUnit,
     val streamFontSize: TextUnit = codeEditorBodyFontSize,
@@ -27,6 +28,7 @@ internal fun regularFont() = AppFont(
     buttonFontSize = 16.sp,
     searchInputSize = 12.sp,
     supplementSize = 11.sp,
+    badgeFontSize = 9.sp,
     codeEditorBodyFontSize = 13.sp,
     codeEditorLineNumberFontSize = 12.sp,